+++ builders/json_serializable_fic/lib/type_helpers.dart
@@ -1,2 +1,8 @@
 export 'package:json_serializable_fic_typehelpers/json_serializable_fic_typehelpers.dart'
-    show FICIListTypeHelper, FICIMapTypeHelper, FICISetTypeHelper;
+    show
+        FICIListTypeHelper,
+        FICIMapTypeHelper,
+        FICISetTypeHelper,
+        FICListSetTypeHelper,
+        FICListMapTypeHelper,
+        FICIMapOfSetsTypeHelper;