+++ typehelpers/json_serializable_fic_typehelpers/lib/json_serializable_fic_typehelpers.dart
@@ -1,2 +1,8 @@
 export 'src/fic_type_helpers.dart'
-    show FICIMapTypeHelper, FICIListTypeHelper, FICISetTypeHelper;
+    show
+        FICIMapTypeHelper,
+        FICIListTypeHelper,
+        FICISetTypeHelper,
+        FICListSetTypeHelper,
+        FICListMapTypeHelper,
+        FICIMapOfSetsTypeHelper;